add a dir for pages' component
