+++ src/pages/DonePage.js
@@ -1,5 +1,4 @@
-import { Container, Row } from "react-bootstrap";
-import { Item } from "./Item";
+
 import { Page } from "./Page";
 import { useTasks } from "../stores";
 import { NoContent } from "../assets/icons";
