+++ src/pages/OverallPage.js
@@ -1,4 +1,4 @@
-import { Container, Row } from "react-bootstrap";
+
 import { useState } from "react";
 import { Page } from "./Page";
 import { NoContent } from "../assets/icons";
