+++ src/pages/page-components/InputAccordion.css
