+++ src/pages/page-components/index.js
@@ -1,0 +1,3 @@
+export {InputAccordion} from './InputAccordion'
+export {Item} from './Item'
+export {Welcome} from './Welcome'
